Add files via upload
Task 1: [30 points] To start with, create this UI. You can set the internal fields of the things like buttons hit or missed to some constant value. This likely will involve a HBOX for your controls and displays, a gridpane for the buttons to be whacked, and a BorderPane to hold those panes.

Task 2: [50 points] Now for this part we want to add functionality. I want you to, on creation, to have one button highlighted in a special way. You can do this by color, an image, or a text label, however you want. When a user clicks on it, you should then highlight a different button. When they click on that, you should highlight another button. However, if they take more than 2 seconds to click on the next button (or the first button) then it should count as a “miss”. When the user clicks the rest button, all the scores should be reset and the game should start anew.


Task 3 [10] : Lastly I want you to gradually “increase” the speed of the game as time goes on. In this case I want you to add an additional display for “High score” at the bottom of the game. This will indicate the most amount of time that a person has went without  “missing” a mole. I want this to be stored between games, so this should be stored in a file. If that file does not exist, then the high score would be 0. However, if I play the game and get a particular score, I want it to be preserved between games in this file